added entitycreator, fixed ECS, added Textureloader

+++ PlayerComponent.h
@@ -2,5 +2,5 @@
 #include <cstdint>
 struct PlayerComponent
 {
-	uint8_t id;
+	unsigned int id;
 };
